+++ lib/gg_git.dart
@@ -9,5 +9,5 @@ library;
 export 'src/gg_git.dart';
 export 'src/commands/get_tags.dart';
 export 'src/commands/commited.dart';
-export 'src/commands/is_pushed.dart';
+export 'src/commands/pushed.dart';
 export 'src/base/gg_git_base.dart';